requirements.txt
telethon==1.34.0
pyyaml==6.0.1
google
groq
pandas
openpyxl
google-genai
# azure-functions
# requests
# lxml
# python-dateutil